Provide a more generous resource restrictions file
The old one was particularly short on allowed ports for networking
calls. This wasn't a problem for RepyV1, where source ports were
automatically chosen by `emulcomm` and it was not required to allow
them explicitly. RepyV2 requires source ports to be in the allowed
list, and so do the RepyV2 unit tests.